feat: Rename or refactor directories/architectures to make it comprehensible for humans.
