+++ src/components/floating-particles/index.ts
